move code closer to where it's used
