+++ internal/graph/graph.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"strings"
-	"sync/atomic"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
@@ -40,38 +39,6 @@ func ContextWithResolutionDepth(parent context.Context, depth uint32) context.Co
 func ResolutionDepthFromContext(ctx context.Context) (uint32, bool) {
 	depth, ok := ctx.Value(resolutionDepthCtxKey).(uint32)
 	return depth, ok
-}
-
-type ResolveCheckRequestMetadata struct {
-	// Thinking of a Check as a tree of evaluations,
-	// Depth is the current level in the tree in the current path that we are exploring.
-	// When we jump one level, we decrement 1. If it hits 0, we throw ErrResolutionDepthExceeded.
-	Depth uint32
-
-	// DispatchCounter is the address to a shared counter that keeps track of how many calls to ResolveCheck we had to do
-	// to solve the root/parent problem.
-	// The contents of this counter will be written by concurrent goroutines.
-	// After the root problem has been solved, this value can be read.
-	DispatchCounter *atomic.Uint32
-
-	// WasThrottled indicates whether the request was throttled
-	WasThrottled *atomic.Bool
-}
-
-func NewCheckRequestMetadata(maxDepth uint32) *ResolveCheckRequestMetadata {
-	return &ResolveCheckRequestMetadata{
-		Depth:           maxDepth,
-		DispatchCounter: new(atomic.Uint32),
-		WasThrottled:    new(atomic.Bool),
-	}
-}
-
-type ResolveCheckResponseMetadata struct {
-	// Number of Read operations accumulated after this request completes.
-	DatastoreQueryCount uint32
-	// Indicates if the ResolveCheck subproblem that was evaluated involved
-	// a cycle in the evaluation.
-	CycleDetected bool
 }
 
 type RelationshipEdgeType int
